// src/types/post.ts
export interface ISocial {
  id: string;
  body?: string | null;
  ownerUserId: string;
  audience?: Audience | null;
  status?: PostStatus | null;
  isDeleted?: boolean | null;
  placeId?: string | null;
  placeDescription?: string | null;
  placeMainText?: string | null;
  placeSecondaryText?: string | null;
  location?: string | null;
  tagIds?: Array<string> | null;
  mentionedUserIds?: Array<string> | null;
  isLikedByUser?: boolean | null;
  isSharedSocialPost?: boolean | null;
  isSharedCampaignPost?: boolean | null;
  userType?: UserTypeEnum | null;
  isMine?: boolean | null;
  videoUrls?: Array<string | null> | null;
  pictureUrls?: Array<string | null> | null;
  createdDateTime?: string | null;
  firstName?: string | null;
  fullName?: string | null;
  lastName?: string | null;
  userName?: string | null;
  userAvatarUrl?: string | null;
  countOfComments?: string | null;
  countOfLikes?: string | null;
  countOfShared?: string | null;
  countOfViews?: string | null;
  postLikerUsers?: Array<{
    id?: string | null;
    avatarUrl?: string | null;
    cognitoUserId?: string | null;
    fullName?: string | null;
    userName?: string | null;
  } | null> | null;
  mentionedUsers?: Array<{
    id?: string | null;
    avatarUrl?: string | null;
    cognitoUserId?: string | null;
    fullName?: string | null;
    userName?: string | null;
  } | null> | null;
  tags?: Array<{ id?: string | null; title?: string | null } | null> | null;
  mediaUrls?: Array<{ isVideo?: boolean | null; url?: string | null } | null> | null;
}
export interface ICampaign {
  id: string;
  body?: string | null;
  title?: string | null;
  ownerUserId: string;
  audience?: Audience | null;
  status?: PostStatus | null;
  isDeleted?: boolean | null;
  placeId?: string | null;
  placeDescription?: string | null;
  target?: any | null;
  deadline?: any | null;
  coverImageUrl?: string | null;
  placeMainText?: string | null;
  placeSecondaryText?: string | null;
  isLikedByUser?: boolean | null;
  userType?: UserTypeEnum | null;
  isMine?: boolean | null;
  location?: string | null;
  tagIds?: Array<string> | null;
  category?: any | null;
  mentionedUserIds?: Array<string> | null;
  videoUrls?: Array<string | null> | null;
  pictureUrls?: Array<string | null> | null;
  createdDateTime?: string | null;
  firstName?: string | null;
  fullName?: string | null;
  lastName?: string | null;
  userName?: string | null;
  userAvatarUrl?: string | null;
  averageRate?: any | null;
  numberOfRates?: string | null;
  raisedMoney?: any | null;
  dayLeft?: number | null;
  numberOfDonations?: string | null;
  countOfComments?: string | null;
  countOfLikes?: string | null;
  countOfShared?: string | null;
  countOfViews?: string | null;
  postLikerUsers?: Array<{
    id?: string | null;
    avatarUrl?: string | null;
    cognitoUserId?: string | null;
    fullName?: string | null;
    userName?: string | null;
  } | null> | null;
  mentionedUsers?: Array<{
    id?: string | null;
    avatarUrl?: string | null;
    cognitoUserId?: string | null;
    fullName?: string | null;
    userName?: string | null;
  } | null> | null;
  tags?: Array<{ id?: string | null; title?: string | null } | null> | null;
  donors?: Array<{
    fullName?: string | null;
    subtitle?: string | null;
    avatarUrl?: string | null;
    donateDate?: string | null;
  } | null> | null;
}

export interface IPost {
  campaign?: ICampaign;
  social?: ISocial;
  share?: ISocial;
}